requirements.txt
requests>=2.28.0
python-dotenv>=0.21.0
scipy>=1.10.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
